+++ src/domain/testUtils/mockPositionInfo.ts
@@ -69,6 +69,7 @@ export function createMockPositionInfo({
     traderDiscountAmount: 0n,
     uiFeeAmount: 0n,
     pendingImpactAmount: 0n,
+    positionValueInUsd: 0n,
     data: "",
 
     marketInfo,