+++ src/lib/chains/blockExplorers.tsx
@@ -6,7 +6,6 @@ import {
   ARBITRUM_SEPOLIA,
   AVALANCHE,
   AVALANCHE_FUJI,
-  BOTANIX,
   MEGAETH,
   getExplorerUrl,
   SOURCE_ETHEREUM_MAINNET,
@@ -28,7 +27,6 @@ export const CHAIN_ID_TO_TX_URL_BUILDER: Record<
   [ARBITRUM_SEPOLIA]: (txId: string) => `${getExplorerUrl(ARBITRUM_SEPOLIA)}tx/${txId}`,
   [AVALANCHE_FUJI]: (txId: string) => `${getExplorerUrl(AVALANCHE_FUJI)}tx/${txId}`,
   [SOURCE_SEPOLIA]: (txId: string) => `${getExplorerUrl(SOURCE_SEPOLIA)}tx/${txId}`,
-  [BOTANIX]: (txId: string) => `${getExplorerUrl(BOTANIX)}tx/${txId}`,
   [MEGAETH]: (txId: string) => `${getExplorerUrl(MEGAETH)}tx/${txId}`,
   [SOURCE_BSC_MAINNET]: (txId: string) => `${getExplorerUrl(SOURCE_BSC_MAINNET)}tx/${txId}`,
   layerzero: (txId: string) => `${getExplorerUrl("layerzero")}tx/${txId}`,
@@ -44,7 +42,6 @@ export const CHAIN_ID_TO_EXPLORER_NAME: Record<AnyChainId, string> = {
   [SOURCE_BASE_MAINNET]: base.blockExplorers.default.name,
   [SOURCE_OPTIMISM_SEPOLIA]: "OP Sepolia Etherscan",
   [SOURCE_SEPOLIA]: "Sepolia Etherscan",
-  [BOTANIX]: "Botanix Explorer",
   [MEGAETH]: "MegaExplorer",
   [SOURCE_BSC_MAINNET]: "BSC Scan",
 };